Implemented MVVM
Add DogListViewModel
Add DogRepository
Update Plugins and IDE
