+++ app/build.gradle
@@ -38,6 +38,29 @@ android {
 
 dependencies {
 
+    //ViewModelScope
+    implementation 'androidx.lifecycle:lifecycle-viewmodel-ktx:2.4.1'
+    //viewModels
+    implementation 'androidx.activity:activity-ktx:1.4.0'
+
+    //Room
+
+    //Retrofit
+
+    //Moshi
+
+    //Picasso
+
+    //Glide
+
+    //Anko
+
+    //View/Layout
+
+    //Coroutines
+
+    //WorkManager
+
     implementation 'androidx.core:core-ktx:1.8.0'
     implementation 'androidx.appcompat:appcompat:1.4.2'
     implementation 'com.google.android.material:material:1.6.1'
