+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Theme.PerroDex"
         tools:targetApi="31">
         <activity
-            android:name=".DogListActivity"
+            android:name=".doglist.DogListActivity"
             android:exported="true">
 
             <intent-filter>
