--- app/src/main/java/com/example/perrodex/DogListActivity.kt
@@ -1,72 +1,0 @@
-package com.example.perrodex
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import androidx.recyclerview.widget.LinearLayoutManager
-import com.example.perrodex.databinding.ActivityDogListBinding
-
-class DogListActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        val binding = ActivityDogListBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-
-        val dogList = getFakeDogs()
-
-        val recycler = binding.dogRecycler
-        recycler.layoutManager = LinearLayoutManager(this)
-
-        val adapter = DogAdapter()
-        recycler.adapter = adapter
-
-
-        adapter.submitList(dogList)
-    }
-
-private fun getFakeDogs(): MutableList<Dog> {
-    val dogList = mutableListOf<Dog>()
-    dogList.add(
-        Dog(
-            1, 1, "Chihuahua", "Toy", 5.4,
-            6.7, "", "12 - 15", "", 10.5,
-            12.3
-        )
-    )
-    dogList.add(
-        Dog(
-            2, 1, "Labrador", "Toy", 5.4,
-            6.7, "", "12 - 15", "", 10.5,
-            12.3
-        )
-    )
-    dogList.add(
-        Dog(
-            3, 1, "Retriever", "Toy", 5.4,
-            6.7, "", "12 - 15", "", 10.5,
-            12.3
-        )
-    )
-    dogList.add(
-        Dog(
-            4, 1, "San Bernardo", "Toy", 5.4,
-            6.7, "", "12 - 15", "", 10.5,
-            12.3
-        )
-    )
-    dogList.add(
-        Dog(
-            5, 1, "Husky", "Toy", 5.4,
-            6.7, "", "12 - 15", "", 10.5,
-            12.3
-        )
-    )
-    dogList.add(
-        Dog(
-            6, 1, "Xoloscuincle", "Toy", 5.4,
-            6.7, "", "12 - 15", "", 10.5,
-            12.3
-        )
-    )
-    return dogList
-}
-}
