+++ app/src/main/java/com/example/perrodex/doglist/DogRepository.kt
@@ -1,0 +1,63 @@
+package com.example.perrodex.doglist
+
+import com.example.perrodex.Dog
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.withContext
+
+class DogRepository {
+
+    //Suspend ya que esta dentro de una corrutina, los recupera y los devuelve al viewModel - MutableLiveData
+    suspend fun downloadDogs(): List<Dog>{
+        //Descargar los datos desde una corrutina - IO (hilo secundario) indica en que ambito se va a utilizar la corrutina -> BBDD etc. - MAIN se ejecuta en el hilo principal
+        return withContext(Dispatchers.IO){
+            getFakeDogs()
+        }
+    }
+
+    private fun getFakeDogs(): MutableList<Dog> {
+        val dogList = mutableListOf<Dog>()
+        dogList.add(
+            Dog(
+                1, 1, "Chihuahua", "Toy", 5.4,
+                6.7, "", "12 - 15", "", 10.5,
+                12.3
+            )
+        )
+        dogList.add(
+            Dog(
+                2, 1, "Labrador", "Toy", 5.4,
+                6.7, "", "12 - 15", "", 10.5,
+                12.3
+            )
+        )
+        dogList.add(
+            Dog(
+                3, 1, "Retriever", "Toy", 5.4,
+                6.7, "", "12 - 15", "", 10.5,
+                12.3
+            )
+        )
+        dogList.add(
+            Dog(
+                4, 1, "San Bernardo", "Toy", 5.4,
+                6.7, "", "12 - 15", "", 10.5,
+                12.3
+            )
+        )
+        dogList.add(
+            Dog(
+                5, 1, "Husky", "Toy", 5.4,
+                6.7, "", "12 - 15", "", 10.5,
+                12.3
+            )
+        )
+        dogList.add(
+            Dog(
+                6, 1, "Xoloscuincle", "Toy", 5.4,
+                6.7, "", "12 - 15", "", 10.5,
+                12.3
+            )
+        )
+        return dogList
+    }
+}
