+++ app/src/main/res/layout/activity_dog_list.xml
@@ -10,7 +10,7 @@
         android:id="@+id/dog_recycler"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".DogListActivity">
+        tools:context=".doglist.DogListActivity">
 
     </androidx.recyclerview.widget.RecyclerView>
 </layout>
