+++ build.gradle
@@ -1,6 +1,6 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 plugins {
-    id 'com.android.application' version '7.4.0-alpha03' apply false
-    id 'com.android.library' version '7.4.0-alpha03' apply false
+    id 'com.android.application' version '7.4.0-alpha05' apply false
+    id 'com.android.library' version '7.4.0-alpha05' apply false
     id 'org.jetbrains.kotlin.android' version '1.6.21' apply false
 }
